+++ src/main/java/webserver/HttpRequest.java
@@ -1,7 +1,7 @@
 package webserver;
 
-import http.util.HttpMessage;
 import http.util.IOUtils;
+import http.util.constant.HttpHeader;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -9,7 +9,17 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class HttpRequest {
-    public static HttpMessage createHttpMessage(BufferedReader br) {
+    private final String[] startLineParam;
+    private final HashMap<String, String> header;
+    private final String body;
+
+    private HttpRequest(String[] startLineParam, HashMap<String, String> header, String body) {
+        this.startLineParam = startLineParam;
+        this.header = header;
+        this.body = body;
+    }
+
+    public static HttpRequest from(BufferedReader br) {
         try {
             String[] startLine = getStartLine(br);
             HashMap<String, String> header = (HashMap<String, String>) parseHeader(br);
@@ -19,9 +29,9 @@ public class HttpRequest {
                 queryString = IOUtils.readData(br, requestContentLength);
             }
 
-            return new HttpMessage(startLine, header, queryString);
+            return new HttpRequest(startLine, header, queryString);
         } catch (Exception e) {
-            return new HttpMessage(new String[0], new HashMap<>(), "");
+            return new HttpRequest(new String[0], new HashMap<>(), "");
         }
     }
 
@@ -50,5 +60,36 @@ public class HttpRequest {
         } catch (Exception e) {
             return new HashMap<>();
         }
+    }
+
+    public String getMethod() {
+        return startLineParam[0];
+    }
+
+    public String getUrl() {
+        return startLineParam[1];
+    }
+
+    public String getCookie() {
+        return header.get(HttpHeader.COOKIE.getHeader());
+    }
+
+    public String getBody() {
+        return body;
+    }
+
+    public static String getContentType(String filePath) {
+        int index = filePath.lastIndexOf(".");
+        String extension = filePath.substring(index + 1);
+
+        if (extension.equals("html") || extension.equals("css") || extension.equals("js")) {
+            return "text/" + extension;
+        }
+
+        if (extension.equals("jpeg") || extension.equals("png")) {
+            return "image/" + extension;
+        }
+
+        return "";
     }
 }
